+++ tests/e2e/builder/cases/polyfill/index.test.ts
@@ -1,7 +1,7 @@
 import path from 'path';
 import { expect, test } from '@modern-js/e2e/playwright';
 import { webpackOnlyTest } from '@scripts/helper';
-import { build } from '@scripts/shared';
+import { build, getHrefByEntryName } from '@scripts/shared';
 
 const POLYFILL_RE = /\/lib-polyfill/;
 
@@ -20,7 +20,9 @@ const getPolyfillContent = (files: Record<string, string>) => {
   return content;
 };
 
-test('should add polyfill when set polyfill entry (default)', async () => {
+test('should add polyfill when set polyfill entry (default)', async ({
+  page,
+}) => {
   const builder = await build({
     cwd: __dirname,
     entry: { index: path.resolve(__dirname, './src/index.js') },
@@ -29,7 +31,15 @@ test('should add polyfill when set polyfill entry (default)', async () => {
         polyfill: 'entry',
       },
     },
+    runServer: true,
   });
+
+  await page.goto(getHrefByEntryName('index', builder.port));
+
+  expect(await page.evaluate('window.a')).toEqual([1, 2, 3, 4, 5, 6, [7, 8]]);
+
+  builder.close();
+
   const files = await builder.unwrapOutputJSON(false);
 
   const content = getPolyfillContent(files);
@@ -39,21 +49,32 @@ test('should add polyfill when set polyfill entry (default)', async () => {
 });
 
 // TODO: needs builtin:swc-loader
-webpackOnlyTest('should add polyfill when set polyfill usage', async () => {
-  const builder = await build({
-    cwd: __dirname,
-    entry: { index: path.resolve(__dirname, './src/index.js') },
-    builderConfig: {
-      output: {
-        polyfill: 'usage',
+webpackOnlyTest(
+  'should add polyfill when set polyfill usage',
+  async ({ page }) => {
+    const builder = await build({
+      cwd: __dirname,
+      entry: { index: path.resolve(__dirname, './src/index.js') },
+      builderConfig: {
+        output: {
+          polyfill: 'usage',
+        },
       },
-    },
-  });
-  const files = await builder.unwrapOutputJSON(false);
+      runServer: true,
+    });
 
-  const content = getPolyfillContent(files);
+    await page.goto(getHrefByEntryName('index', builder.port));
 
-  // should only polyfill some usage api
-  expect(content.includes('es.array.flat.js')).toBeTruthy();
-  expect(content.includes('String.prototype.trimEnd')).toBeFalsy();
-});
+    expect(await page.evaluate('window.a')).toEqual([1, 2, 3, 4, 5, 6, [7, 8]]);
+
+    builder.close();
+
+    const files = await builder.unwrapOutputJSON(false);
+
+    const content = getPolyfillContent(files);
+
+    // should only polyfill some usage api
+    expect(content.includes('es.array.flat.js')).toBeTruthy();
+    expect(content.includes('String.prototype.trimEnd')).toBeFalsy();
+  },
+);