+++ tests/integration/basic-app/modern.config.ts
@@ -1,0 +1,7 @@
+import { applyBaseConfig } from '../../utils/applyBaseConfig';
+
+export default applyBaseConfig({
+  runtime: {
+    state: true,
+  },
+});