+++ tests/integration/ssg/fixtures/nested-routes/modern.config.ts
@@ -1,4 +1,4 @@
-import { appTools, defineConfig } from '@modern-js/app-tools';
+import { defineConfig, appTools } from '@modern-js/app-tools';
 import { ssgPlugin } from '@modern-js/plugin-ssg';
 
 export default defineConfig({
@@ -7,6 +7,13 @@ export default defineConfig({
   },
   output: {
     ssg: true,
+    polyfill: 'off',
+    disableTsChecker: true,
+  },
+  performance: {
+    chunkSplit: {
+      strategy: 'all-in-one',
+    },
   },
   plugins: [appTools(), ssgPlugin()],
 });