+++ tests/integration/tailwindcss/fixtures/tailwindcss-v3-js-config/modern.config.ts
@@ -1,14 +1,6 @@
-import { appTools, defineConfig } from '@modern-js/app-tools';
 import { tailwindcssPlugin } from '@modern-js/plugin-tailwindcss';
+import { applyBaseConfig } from '../../../../utils/applyBaseConfig';
 
-export default defineConfig({
-  plugins: [
-    appTools({
-      bundler:
-        process.env.PROVIDE_TYPE === 'rspack'
-          ? 'experimental-rspack'
-          : 'webpack',
-    }),
-    tailwindcssPlugin(),
-  ],
+export default applyBaseConfig({
+  plugins: [tailwindcssPlugin()],
 });