+++ tests/integration/write-to-dist/modern.config.ts
@@ -1,11 +1,8 @@
-import { appTools, defineConfig } from '@modern-js/app-tools';
+import { applyBaseConfig } from '../../utils/applyBaseConfig';
 
-export default defineConfig({
+export default applyBaseConfig({
   runtime: {
     router: true,
-  },
-  server: {
-    // ssr: true,
   },
   tools: {
     devServer: {
@@ -14,12 +11,4 @@ export default defineConfig({
       },
     },
   },
-  plugins: [
-    appTools({
-      bundler:
-        process.env.PROVIDE_TYPE === 'rspack'
-          ? 'experimental-rspack'
-          : 'webpack',
-    }),
-  ],
 });