+++ packages/toolkit/compiler/babel/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @modern-js/babel-compiler
 
+## 2.35.0
+
+### Patch Changes
+
+- Updated dependencies [15b834f]
+  - @modern-js/utils@2.35.0
+
 ## 2.34.0
 
 ### Patch Changes